requirements.txt
# Core dependencies for Automated Fault Tracing
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyYAML==6.0.1
python-multipart==0.0.6
